requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.10
black==25.1.0
click==8.2.1
dill==0.4.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
isort==6.0.1
mako==1.3.10
markupsafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic-core==2.33.2
pyfiglet==1.0.2
pylint==3.3.7
python-dotenv==1.1.0
sniffio==1.3.1
sqlalchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
tomlkit==0.13.2
typing-extensions==4.13.2
typing-inspection==0.4.1
uvicorn==0.34.2
watchfiles==1.0.5
